gcode scripts ui
