+++ client_src/js/lib/modules/fileManager.js
@@ -12,6 +12,7 @@ import {
   generateTableRows,
   updateTableRow
 } from "../../printer-manager/functions/bulk-actions-progress.functions";
+import gcodeScripts from "../../services/gcode-scripts.service";
 
 const fileUploads = new Queue();
